Improve documentation and rename accordingly
