+++ GDEdit/GDE.App/Main/Screens/Edit/Editor.cs
@@ -7,7 +7,7 @@ using osu.Framework.Graphics;
 using osu.Framework.Screens;
 using osuTK;
 using osuTK.Graphics;
-using GDE.App.Main.Colours;
+using GDE.App.Main.Colors;
 using GDE.App.Main.UI;
 using GDE.App.Main.UI.Toolbar;
 using GDEdit.Application;
