+++ GDEdit/GDE.App/Main/Screens/LevelCreation.cs
@@ -7,7 +7,7 @@ using osu.Framework.Graphics;
 using osu.Framework.Screens;
 using osuTK;
 using osuTK.Graphics;
-using GDE.App.Main.Colours;
+using GDE.App.Main.Colors;
 using GDE.App.Main.UI;
 using GDEdit.Application;
 using GDEdit.Application.Editor;
@@ -32,7 +32,7 @@ namespace GDE.App.Main.Screens
                     RelativeSizeAxes = Axes.Both,
                     Anchor = Anchor.Centre,
                     Origin = Anchor.Centre,
-                    Colour = GDEColours.FromHex("111")
+                    Colour = GDEColors.FromHex("111")
                 },
                 name = new TextBox
                 {
